+++ trunk/server/src/server/spell_effect.c
@@ -1317,7 +1317,7 @@ int cast_change_attr(object *op, object *caster, object *target, int dir, int sp
     {
         if (tmp2->type == FORCE)
         {
-            if (tmp2->weight_limit == (uint32) spell_type)
+            if (tmp2->weight_limit == spell_type)
             {
                 force = tmp2;    /* the old effect will be "refreshed" */
                 is_refresh = 1;
@@ -1745,9 +1745,9 @@ int summon_pet(object *op, int dir, SpellTypeFrom item)
     LOG(llevBug, "BUG: unimplemented summon pet spell used\n");
     return 0;
     /* TODO: update for the new AI and pet systems. Gecko 2006-04-30 */
-#if 0        
-    
-    
+#if 0
+
+
     int         level, number, i;
     char       *monster;
     archetype  *at;
@@ -3260,7 +3260,7 @@ int cast_charm(object *op, object *caster, archetype *arch, int spellnum)
     LOG(llevBug, "BUG: unimplemented spell 'charm'\n");
     return 0;
     /* TODO: update for the new AI and pet systems. Gecko 2006-04-30 */
-#if 0        
+#if 0
     int         i, xt, yt;
     object     *tmp, *effect;
     mapstruct  *m;
@@ -3300,7 +3300,7 @@ int cast_charm(object *op, object *caster, archetype *arch, int spellnum)
         tmp->move_type = PETMOVE;
     }
     return 1;
-#endif    
+#endif
 }
 
 int cast_charm_undead(object *op, object *caster, archetype *arch, int spellnum)
@@ -3308,7 +3308,7 @@ int cast_charm_undead(object *op, object *caster, archetype *arch, int spellnum)
     LOG(llevBug, "BUG: unimplemented spell 'charm undead'\n");
     return 0;
     /* TODO: update for the new AI and pet systems. Gecko 2006-04-30 */
-#if 0        
+#if 0
     int         i, bonus, xt, yt;
     mapstruct  *m;
     object     *tmp, *effect, *god = find_god(determine_god(op));
@@ -3359,7 +3359,7 @@ int cast_charm_undead(object *op, object *caster, archetype *arch, int spellnum)
         tmp->move_type = PETMOVE;
     }
     return 1;
-#endif    
+#endif
 }
 
 /* Returns a monster (chosen at random) that this particular player (and his
@@ -3694,7 +3694,7 @@ object * fix_summon_pet(archetype *at, object *op, int dir, int type)
     LOG(llevBug, "BUG: unimplemented function 'fix_summon_pet()' called\n");
     return 0;
     /* TODO: update for the new AI and pet systems. Gecko 2006-04-30 */
-#if 0        
+#if 0
     archetype      *atmp;
     object*tmp =    NULL, *prev = NULL, *head = NULL;
 
@@ -3769,7 +3769,7 @@ object * fix_summon_pet(archetype *at, object *op, int dir, int type)
         CLEAR_FLAG(head, FLAG_SPLITTING);
 
     return head;
-#endif    
+#endif
 }
 
 /* cast_consecrate() - a spell to make an altar your god's */
@@ -3880,7 +3880,7 @@ int animate_weapon(object *op, object *caster, int dir, archetype *at, int spell
     LOG(llevBug, "BUG: unimplemented spell 'animate_weapon'\n");
     return 0;
     /* TODO: update for the new AI and pet systems. Gecko 2006-04-30 */
-#if 0        
+#if 0
     object *weapon, *tmp;
     char    buf[MAX_BUF];
     int     a, i, j;
@@ -4131,7 +4131,7 @@ int animate_weapon(object *op, object *caster, int dir, archetype *at, int spell
     tmp->direction = dir;
     insert_ob_in_map(tmp, op->map, op, 0);
     return 1;
-#endif    
+#endif
 }
 
 
@@ -4407,7 +4407,7 @@ void move_peacemaker(object *op)
 {
     LOG(llevBug, "BUG: unimplemented function 'move_peacemaker'\n");
     /* TODO: update for the new AI and pet systems. Gecko 2006-04-30 */
-#if 0        
+#if 0
     object *tmp;
     char    buf[MAX_BUF];
 
@@ -4447,7 +4447,7 @@ void move_peacemaker(object *op)
             }
         }
     }
-#endif    
+#endif
 }
 
 
@@ -4456,7 +4456,7 @@ int cast_cause_conflict(object *op, object *caster, archetype *spellarch, int ty
     LOG(llevBug, "BUG: unimplemented spell 'casuse conflict'\n");
     return 0;
     /* TODO: update for the new AI and pet systems. Gecko 2006-04-30 */
-#if 0        
+#if 0
     int         i, j, xt, yt;
     int         r;    /*  peterm:  added to make area of effect level dep.  */
     int         level;
@@ -4496,6 +4496,6 @@ int cast_cause_conflict(object *op, object *caster, archetype *spellarch, int ty
             }
         }
     return 1;
-#endif    
+#endif
 }
